Refactor Rust module extraction and parsing
- Replace the `extraction` module with a new `parsing` module for improved clarity in module extraction logic.
- Update the `RustAnalyser` to utilize the new `parsing::extract_modules_from_file` function, enhancing the organization of code.
- Introduce the `parsing.rs` file, which contains the logic for extracting modules and their associated symbols from Rust source files, including handling of visibility and documentation comments.
- Adjust related tests to ensure accurate extraction and representation of modules and symbols, maintaining consistency with the new parsing approach.
